refactor folder structure to mirror firestore db structure
