+++ functions/firestore/movieClubs/movies/comments/commentFunctions.js
@@ -1,0 +1,23 @@
+const { db } = require('firestore');
+
+async function postComment(movieClubId, movieId, commentData) {
+    try {
+        // Reference to the specific movie's comments collection
+        const commentsRef = db
+            .collection('movieclubs')
+            .doc(movieClubId)
+            .collection('movies')
+            .doc(movieId)
+            .collection('comments');
+
+        // Add the comment to the 'comments' sub-collection
+        await commentsRef.add({
+            ...commentData,
+            createdAt: admin.firestore.FieldValue.serverTimestamp(),
+        });
+
+        console.log('Comment posted successfully!');
+    } catch (error) {
+        console.error('Error posting comment:', error);
+    }
+}
